#36 update: 편의시설들 컨트롤러 하나로 통합 편의시설 컨트롤러 5개 하나로 병합
