+++ backend/src/main/java/com/whereismyhome/amenities/dto/BusResponseDto.java
@@ -1,0 +1,19 @@
+package com.whereismyhome.amenities.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class BusResponseDto {
+
+    private Integer id;
+
+    private String number;
+
+    private String name;
+
+    private String lat;
+
+    private String lng;
+}
