+++ backend/src/main/java/com/whereismyhome/amenities/dto/HospitalResponseDto.java
@@ -1,0 +1,24 @@
+package com.whereismyhome.amenities.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class HospitalResponseDto {
+
+    private Integer id;
+
+
+    private String name;
+
+
+    private String phoneNumber;
+
+
+    private String lng;
+
+
+    private String lat;
+}
+
