+++ backend/src/main/java/com/whereismyhome/amenities/dto/SchoolResponseDto.java
@@ -1,0 +1,16 @@
+package com.whereismyhome.amenities.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class SchoolResponseDto {
+    private Integer id;
+
+    private String name;
+
+    private String lat;
+
+    private String lng;
+}
