+++ backend/src/main/java/com/whereismyhome/amenities/dto/SubwayResponseDto.java
@@ -1,0 +1,19 @@
+package com.whereismyhome.amenities.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class SubwayResponseDto {
+
+    private Integer id;
+
+    private String name;
+
+    private String lat;
+
+    private String lng;
+
+    private String phoneNumber;
+}
