+++ backend/src/main/java/com/whereismyhome/amenities/entity/Bus.java
@@ -1,0 +1,30 @@
+package com.whereismyhome.amenities.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity(name = "bus")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Bus {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @Column
+    private String number;
+
+    @Column
+    private String name;
+
+    @Column
+    private String lat;
+
+    @Column
+    private String lng;
+}
