+++ backend/src/main/java/com/whereismyhome/amenities/entity/School.java
@@ -1,0 +1,27 @@
+package com.whereismyhome.amenities.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity(name = "school")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class School {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @Column
+    private String name;
+
+    @Column
+    private String lat;
+
+    @Column
+    private String lng;
+}
