+++ backend/src/main/java/com/whereismyhome/amenities/repository/BusRepository.java
@@ -1,0 +1,7 @@
+package com.whereismyhome.amenities.repository;
+
+import com.whereismyhome.amenities.entity.Bus;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BusRepository extends JpaRepository<Bus,Integer> {
+}
