+++ backend/src/main/java/com/whereismyhome/amenities/repository/CctvRepository.java
@@ -1,0 +1,7 @@
+package com.whereismyhome.amenities.repository;
+
+import com.whereismyhome.amenities.entity.Cctv;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CctvRepository extends JpaRepository<Cctv,Integer> {
+}
