+++ backend/src/main/java/com/whereismyhome/amenities/repository/SchoolRepository.java
@@ -1,0 +1,7 @@
+package com.whereismyhome.amenities.repository;
+
+import com.whereismyhome.amenities.entity.School;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SchoolRepository extends JpaRepository<School,Integer> {
+}
