+++ backend/src/main/java/com/whereismyhome/amenities/repository/SubwayRepository.java
@@ -1,0 +1,7 @@
+package com.whereismyhome.amenities.repository;
+
+import com.whereismyhome.amenities.entity.Subway;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SubwayRepository extends JpaRepository<Subway,Integer> {
+}
