--- backend/src/main/java/com/whereismyhome/bus/entity/Bus.java
@@ -1,30 +1,0 @@
-package com.whereismyhome.bus.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity(name = "bus")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Bus {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column
-    private String number;
-
-    @Column
-    private String name;
-
-    @Column
-    private String lat;
-
-    @Column
-    private String lng;
-}
