--- backend/src/main/java/com/whereismyhome/bus/repository/BusRepository.java
@@ -1,7 +1,0 @@
-package com.whereismyhome.bus.repository;
-
-import com.whereismyhome.bus.entity.Bus;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BusRepository extends JpaRepository<Bus,Integer> {
-}
