--- backend/src/main/java/com/whereismyhome/cctv/mapper/CctvMapper.java
@@ -1,12 +1,0 @@
-package com.whereismyhome.cctv.mapper;
-
-import com.whereismyhome.cctv.dto.CctvResponseDto;
-import com.whereismyhome.cctv.entity.Cctv;
-import org.mapstruct.Mapper;
-
-import java.util.List;
-
-@Mapper(componentModel = "spring")
-public interface CctvMapper {
-    List<CctvResponseDto> cctvListToCctvResponseDtos(List<Cctv> cctvList);
-}
