--- backend/src/main/java/com/whereismyhome/cctv/repository/CctvRepository.java
@@ -1,7 +1,0 @@
-package com.whereismyhome.cctv.repository;
-
-import com.whereismyhome.cctv.entity.Cctv;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CctvRepository extends JpaRepository<Cctv,Integer> {
-}
