--- backend/src/main/java/com/whereismyhome/hospital/dto/HospitalResponseDto.java
@@ -1,24 +1,0 @@
-package com.whereismyhome.hospital.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class HospitalResponseDto {
-
-    private Integer id;
-
-
-    private String name;
-
-
-    private String phoneNumber;
-
-
-    private String lng;
-
-
-    private String lat;
-}
-
