--- backend/src/main/java/com/whereismyhome/hospital/entity/Hospital.java
@@ -1,31 +1,0 @@
-package com.whereismyhome.hospital.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Hospital {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column
-    private String name;
-
-    @Column
-    private String phoneNumber;
-
-    @Column
-    private String lng;
-
-    @Column
-    private String lat;
-}
