--- backend/src/main/java/com/whereismyhome/hospital/repository/HospitalRepository.java
@@ -1,7 +1,0 @@
-package com.whereismyhome.hospital.repository;
-
-import com.whereismyhome.hospital.entity.Hospital;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface HospitalRepository extends JpaRepository<Hospital, Integer> {
-}
