--- backend/src/main/java/com/whereismyhome/shool/dto/SchoolResponseDto.java
@@ -1,16 +1,0 @@
-package com.whereismyhome.shool.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class SchoolResponseDto {
-    private Integer id;
-
-    private String name;
-
-    private String lat;
-
-    private String lng;
-}
