--- backend/src/main/java/com/whereismyhome/shool/entity/School.java
@@ -1,27 +1,0 @@
-package com.whereismyhome.shool.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity(name = "school")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class School {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column
-    private String name;
-
-    @Column
-    private String lat;
-
-    @Column
-    private String lng;
-}
