--- backend/src/main/java/com/whereismyhome/subway/dto/SubwayResponseDto.java
@@ -1,19 +1,0 @@
-package com.whereismyhome.subway.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class SubwayResponseDto {
-
-    private Integer id;
-
-    private String name;
-
-    private String lat;
-
-    private String lng;
-
-    private String phoneNumber;
-}
