--- backend/src/main/java/com/whereismyhome/subway/entity/Subway.java
@@ -1,30 +1,0 @@
-package com.whereismyhome.subway.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity(name = "subway")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Subway {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column
-    private String name;
-
-    @Column
-    private String lat;
-
-    @Column
-    private String lng;
-
-    @Column
-    private String phoneNumber;
-}
