--- backend/src/main/java/com/whereismyhome/subway/repository/SubwayRepository.java
@@ -1,7 +1,0 @@
-package com.whereismyhome.subway.repository;
-
-import com.whereismyhome.subway.entity.Subway;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SubwayRepository extends JpaRepository<Subway,Integer> {
-}
